# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EDIFICIA API
  version: 1.0.0
  description: |
    API para la redacción automatizada de Memorias de Proyecto de Ejecución (CTE/LOE).
    
    ## Autenticación
    La API usa JWT Bearer tokens. Obtener un token mediante `POST /api/auth/login` 
    e incluirlo en el header `Authorization: Bearer <token>`.
    
    ## Refresh Tokens
    Los access tokens expiran en 60 minutos. Usar `POST /api/auth/refresh` para 
    obtener un nuevo par de tokens (rotation). Se implementa stolen-token detection.
    
    ## Errores
    Todas las respuestas de error usan el formato RFC 7807 ProblemDetails.
  contact:
    name: EDIFICIA
    url: https://edificia.jesusjbriceno.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:5000
    description: Local
  - url: https://api-edificia.jesusjbriceno.dev
    description: Producción
security:
  - BearerAuth: []
...
paths:
...
  /api/projects/{id}/ai/generate:
    post:
      tags: [AI]
      summary: Generate Section Text
      description: |
        Genera texto para una sección del proyecto usando IA (Flux Gateway).
        El backend enriquece el prompt con metadatos del proyecto (título, tipo de intervención,
        LOE, dirección, normativa local, contenido existente de la sección).
      operationId: generateSectionText
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTextRequest'
      responses:
        '200':
          description: Texto generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedTextResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── EXPORT ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token JWT obtenido de POST /api/auth/login. Ejemplo: eyJhbGciOiJIUzI1NiIs...'

  # ─── PARAMETERS ─────────────────────────────────────────
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID del proyecto
...
  responses:
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 401
            title: No autorizado
            detail: Token no válido o expirado.
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 404
            title: No encontrado
            detail: El recurso solicitado no existe.
...
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        status:
          type: integer
          example: 400
        title:
          type: string
          example: Error en la solicitud
        detail:
          type: string
          example: Descripción del error.
        code:
          type: string
          example: Entity.ErrorCode
...
    GenerateTextRequest:
      type: object
      required: [sectionId, prompt]
      properties:
        sectionId:
          type: string
          maxLength: 200
          example: MD.01
          description: ID de la sección dentro del árbol de contenido
        prompt:
          type: string
          maxLength: 10000
          example: Describe los agentes intervinientes en el proyecto
        context:
          type: string
          nullable: true
          maxLength: 50000
          description: Contexto adicional opcional
    GeneratedTextResponse:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        sectionId:
          type: string
          example: MD.01
        generatedText:
          type: string
          description: Texto HTML generado por IA